+++ src/components/FileInput.vue
@@ -1,21 +1,19 @@
 <template>
   <div>
-    <h1>{{ file.name }}</h1>
-    <div v-on:click="$emit('delete')" class="icon-delete">
-      <i class="icon-minus"></i>
-    </div>
-    <br />
-
-    <DisplayImage :file="file" />
+    <Card @delete="$emit('delete')" :file="file" :locked="!/^image/.test(file.meta.mime)">
+      <DisplayImage :file="file" />
+    </Card>
   </div>
 </template>
 
 <script lang="ts">
  import DisplayImage from './DisplayImage.vue'
+ import Card from './Card.vue';
 
  export default {
    components: {
      DisplayImage,
+     Card,
    },
    props: ['file']
  }
@@ -26,11 +24,4 @@
    display: inline-block;
    margin-top: 0;
  }
- .icon-delete {
-   color: red;
-   display: inline;
-   padding: 2px;
-   cursor: pointer;
- }
-
 </style>